flask-server/requirements.txt
allure-pytest==2.9.43
allure-python-commons==2.9.43
appdirs==1.4.4
asn1crypto==1.5.1
atomicwrites==1.4.0
attrs==21.2.0
autopep8==1.5.7
Babel==2.11.0
backports.entry-points-selectable==1.1.0
bcrypt==4.0.1
blinker==1.5
certifi==2021.5.30
cffi==1.15.1
charset-normalizer==2.0.6
click==8.0.1
colorama==0.4.4
cryptography==36.0.2
deepdiff==5.8.0
distlib==0.3.2
dnspython==2.1.0
docker==6.0.1
email-validator==1.1.3
et-xmlfile==1.1.0
exceptiongroup==1.1.0
filelock==3.0.12
flake8==4.0.1
Flask==2.0.1
Flask-BabelEx==0.9.4
Flask-Cors==3.0.10
flask-extended==0.2
Flask-JWT==0.3.2
Flask-JWT-Extended==4.4.0
Flask-Login==0.6.2
Flask-Mail==0.9.1
flask-mongoengine==1.0.0
Flask-Principal==0.4.0
Flask-PyMongo==2.3.0
Flask-Security==3.0.0
flask-swagger-ui==3.36.0
Flask-WTF==0.15.1
idna==3.2
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.2.0
MarkupSafe==2.0.1
mccabe==0.6.1
mongoengine==0.23.1
mysql-connector-python==8.0.28
numpy==1.21.0
openpyxl==3.0.9
ordered-set==4.1.0
oscrypto==1.3.0
packaging==21.0
pandas==1.4.1
passlib==1.7.4
pep8==1.7.1
Pillow==9.4.0
platformdirs==2.0.2
pluggy==0.13.1
protobuf==3.17.3
psutil==5.8.0
py==1.10.0
pycodestyle==2.8.0
pycparser==2.21
pycryptodomex==3.17
pyflakes==2.4.0
PyJWT==2.6.0
pymongo==3.12.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pytest==6.2.4
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.3
pywin32==305
PyYAML==6.0
requests==2.26.0
scikit-learn==1.2.1
scipy==1.10.1
six==1.16.0
snowflake-connector-python==2.7.6
speaklater==1.3
threadpoolctl==3.1.0
toml==0.10.2
tomli==2.0.1
urllib3==1.26.7
virtualenv==20.6.0
websocket-client==1.5.1
Werkzeug==2.0.1
WTForms==2.3.3